Improve the visibility of process entries
Before this change it was impossible to see the command arguments, as the executable was to long.

This changes that so we use the process name as executable name VSCode as seems to do this. So you could still see the arguments of the program.

I also added a tooltip to see the correct executable + arguments.